remove version check for raw_input as input
For compatibility with python3, we want to check if we are using
python < 3, and if so, rename raw_input to input. Instead of
explicitly checking the version, now we simply see if the raw_input
function exists, and if so alias it to input, so that the behavior
of input() is consistent on pthon2 and python 3.